Part 4

+++ osa4/blogilista/controllers/blogs.js
@@ -1,17 +1,28 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user')
     response.json(blogs.map(blog => blog.toJSON()))
   } catch (exception) {
     next(exception)
   }
 })
 
+
 blogsRouter.post('/', async (request, response, next) => {
   try {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!decodedToken || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
     const blog = new Blog(request.body)
     if (!blog.url) {
       return response.status(400).end()
@@ -20,10 +31,16 @@ blogsRouter.post('/', async (request, response, next) => {
       return response.status(400).end()
     }
 
+    blog.user = user.id
+
     if (!blog.likes) {
       blog.likes = 0
     }
     const result = await blog.save()
+    const resJSON = result.toJSON()
+    const id = resJSON.id
+    user.blogs = user.blogs.concat(id)
+    await user.save()
     response.status(201).json(result)
   } catch (exception) {
     next(exception)
@@ -32,8 +49,20 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (decodedToken.id === blog.user.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      return response.status(204).end()
+    }
+    return response.status(401).json({ error: 'can only delete own blog' })
+
+
   } catch (exception) {
     next(exception)
   }
